Add files via upload
This Python script creates a web-based language translation application using Streamlit. It allows users to input text and select a target language for translation. The language data, including names and ISO codes, is loaded from a CSV file (language.csv). A dictionary is created to map language names to their corresponding ISO codes. The app uses the translate package to perform the translation. It features a user-friendly interface with a text area for input and a sidebar for language selection.

Upon entering text, the script translates it into the selected language and displays the translated text. If the selected language is supported by Google Text-to-Speech (gTTS), the script generates an audio file of the translated text. Users can listen to the translation directly on the app and download the audio file as well.

The function get_binary_file_downloader_html enables users to download the generated audio file by encoding it in base64 format. The script also handles potential errors during the translation or text-to-speech process and displays an error message if any issues arise. This app provides a simple, interactive tool for translating text and converting it into speech, supporting a wide range of languages.